refactor(ts): split builtins.ts into builtins/ package (scope/dirs/vars/man/script/condition/text) (mirror #146)
