+++ typescript/packages/core/src/workspace/executor/builtins/scope.ts
@@ -1,0 +1,32 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import type { ByteSource, IOResult as IOResultType } from '../../../io/types.ts'
+import type { IOResult } from '../../../io/types.ts'
+import { PathSpec } from '../../../types.ts'
+import type { ExecutionNode } from '../../types.ts'
+
+export type Result = [ByteSource | null, IOResult, ExecutionNode]
+
+export type ExecuteStringFn = (script: string, opts: { sessionId: string }) => Promise<IOResultType>
+
+export function toScope(path: string): PathSpec {
+  const lastSlash = path.lastIndexOf('/')
+  const directory = lastSlash >= 0 ? path.slice(0, lastSlash + 1) : '/'
+  return new PathSpec({ original: path, directory, resolved: true })
+}
+
+export function scopePath(val: string | PathSpec): string {
+  return val instanceof PathSpec ? val.original : val
+}
